Move code from __master__ to kernel.py so it can be imported
